fix: add composite actions to Dependabot (#8)
This commit fixes an issue where GitHub Actions dependencies in
composite actions are not bumped by Dependabot[1]. To resolve this, the
existing workaround is to specify the paths of the composite actions
explicitly.

As part of this, we also rollback one of the dependencies to an earlier
version so we can validate the fix.

[1]: https://github.com/dependabot/dependabot-core/issues/6704